+++ src/compiler/emitter/mod.rs
@@ -8,5 +8,7 @@ mod lines_2000_2500;
 pub use lines_2000_2500::*;
 mod lines_2500_3000;
 pub use lines_2500_3000::*;
-mod lines_3000_end;
-pub use lines_3000_end::*;
+mod lines_3000_3500;
+pub use lines_3000_3500::*;
+mod lines_3500_end;
+pub use lines_3500_end::*;
